requirements.txt
Django
django-phonenumber-field